remove deprecated artifact resources
